+++ changelog.d/CHANGELOG.md
@@ -60,6 +60,15 @@
   This may have impact if you have multiple unauthenticated SSH connections in a short time.
   We tested this change on non-production machines over the last 3 weeks and got no reports of problems.
 
+# Release 2024_033
+
+## NixOS XX.XX platform
+
+- physical machines: load `dm_mirror` kernel module by default, to support several LVM disk migration scenarios
+
+- Update fc.qemu to ensure reduce cluster load on rbd list. (PL-133194)
+
+
 # Release 2024_032
 
 ## NixOS XX.XX platform